+++ lualib/http/websocket.lua
@@ -9,6 +9,15 @@ local socket_error = sockethelper.socket_error
 local GLOBAL_GUID = "258EAFA5-E914-47DA-95CA-C5AB0DC85B11"
 local MAX_FRAME_SIZE = 256 * 1024 -- max frame is 256K
 
+local assert = assert
+local pairs = pairs
+local error = error
+local string = string
+local xpcall = xpcall
+local debug = debug
+local table = table
+local tonumber = tonumber
+
 local M = {}
 
 
@@ -24,6 +33,27 @@ local function _isws_closed(id)
     return not ws_pool[id]
 end
 
+local function reader_with_payload(self, payload)
+    local sz_payload = #payload
+    if sz_payload == 0 then
+        return
+    end
+    local read = self.read
+    function self.read (sz)
+        if sz == nil or sz == sz_payload then
+            self.read = read
+            return payload
+        end
+        if sz < sz_payload then
+            local ret = payload:sub(1, sz)
+            payload = payload:sub(sz + 1)
+            sz_payload = #payload
+            return ret
+        end
+        self.read = read
+        return payload .. read(sz - sz_payload)
+    end
+end
 
 local function write_handshake(self, host, url, header)
     local key = crypt.base64encode(crypt.randomkey()..crypt.randomkey())
@@ -41,11 +71,11 @@ local function write_handshake(self, host, url, header)
     end
 
     local recvheader = {}
-    local code, body = internal.request(self, "GET", host, url, recvheader, request_header)
+    local code, payload = internal.request(self, "GET", host, url, recvheader, request_header)
     if code ~= 101 then
-        error(string.format("websocket handshake error: code[%s] info:%s", code, body))
+        error(string.format("websocket handshake error: code[%s] info:%s", code, payload))
     end
-	assert(body == "")	-- todo: M.read may need handle it
+    reader_with_payload(self, payload)
 
     if not recvheader["upgrade"] or recvheader["upgrade"]:lower() ~= "websocket" then
         error("websocket handshake upgrade must websocket")
@@ -67,17 +97,17 @@ local function write_handshake(self, host, url, header)
     end
 end
 
-
 local function read_handshake(self, upgrade_ops)
     local header, method, url
     if upgrade_ops then
         header, method, url = upgrade_ops.header, upgrade_ops.method, upgrade_ops.url
     else
         local tmpline = {}
-        local header_body = internal.recvheader(self.read, tmpline, "")
-        if not header_body then
+        local payload = internal.recvheader(self.read, tmpline, "")
+        if not payload then
             return 413
         end
+        reader_with_payload(self, payload)
 
         local request = assert(tmpline[1])
         local httpver
@@ -526,5 +556,6 @@ function M.close(id, code ,reason)
     end
 end
 
+M.is_close = _isws_closed
 
 return M