+++ service-src/service_snlua.c
@@ -233,7 +233,7 @@ static int luaB_auxwrap (lua_State *L) {
   if (r < 0) {
     int stat = lua_status(co);
     if (stat != LUA_OK && stat != LUA_YIELD)
-      lua_resetthread(co, L);  /* close variables in case of errors */
+      lua_closethread(co, L);  /* close variables in case of errors */
     if (lua_type(L, -1) == LUA_TSTRING) {  /* error object is a string? */
       luaL_where(L, 1);  /* add extra info, if available */
       lua_insert(L, -2);