+++ skynet-src/atomic.h
@@ -1,10 +1,10 @@
 #ifndef SKYNET_ATOMIC_H
 #define SKYNET_ATOMIC_H
 
-#ifdef __STDC_NO_ATOMICS__
-
 #include <stddef.h>
 #include <stdint.h>
+
+#ifdef __STDC_NO_ATOMICS__
 
 #define ATOM_INT volatile int
 #define ATOM_POINTER volatile uintptr_t